reorg mask and annotation utils a bit
